Backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
mysqlclient>=2.2,<3.0
django-cors-headers>=4.4,<5.0
django-filter>=24.2,<25.0

